fix the DOMContentLoaded event binding issue when async loading javapoly lib
for #83

when async loading javapoly lib in browser, we may miss the 'DOMContentLoaded' event in the Java poly code.
the DOM content may be already loaded when we async loading javapoly lib.
the only guarantee of the the async script is that it run at some mysterious point before the 'load' event gets fired.

here we just add code check the readyState, if it's already loaded, we will directly start the JVM. otherwise we will install a listener on 'DOMContentLoaded' event to start JVM.